Fix: Upload images in blocks
This also uses two dependencies for managing the threading with the ImageUploadTask. I hope to get all of that rewritten in the next few days to be more streamlined. Perhaps also refactor the BufferUploader interface a bit to open the API up a bit more. Also, this REALLY needs some general code refactoring...